moved protocol out of adapter directory
