+++ jobbers/adapters/redis_json.py
@@ -37,7 +37,7 @@ if TYPE_CHECKING:
     from pydantic import BaseModel
     from redis.asyncio.client import Pipeline, Redis
 
-    from jobbers.adapters.protocols import TaskAdapterProtocol
+    from jobbers.protocols import TaskAdapterProtocol
 
 
 def _escape_tag(value: str) -> str:
